+++ src/main/java/ch/newinstance/plugin/mavendependencychecker/util/QueryBuilder.java
@@ -1,6 +1,7 @@
 package ch.newinstance.plugin.mavendependencychecker.util;
 
 import org.apache.maven.model.Dependency;
+import org.apache.maven.model.Plugin;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -12,8 +13,12 @@ public class QueryBuilder {
     private static final String AND = "+AND+";
     private static final String QUOTATION = "%20";
 
-    public List<String> buildQueries(List<Dependency> dependencies) {
+    public List<String> buildDependencyQueries(List<Dependency> dependencies) {
         return dependencies.stream().map(dependency -> buildQuery(dependency.getGroupId(), dependency.getArtifactId())).collect(Collectors.toList());
+    }
+
+    public List<String> buildPluginQueries(List<Plugin> plugins) {
+        return plugins.stream().map(plugin -> buildQuery(plugin.getGroupId(), plugin.getArtifactId())).collect(Collectors.toList());
     }
 
     private String buildQuery(String groupId, String artifactId) {