+++ src/main/java/ch/newinstance/plugin/mavendependencychecker/util/VersionComparator.java
@@ -4,6 +4,8 @@ import ch.newinstance.plugin.mavendependencychecker.model.DependencyUpdateResult
 
 import org.apache.maven.artifact.versioning.ComparableVersion;
 import org.apache.maven.model.Dependency;
+import org.apache.maven.model.Plugin;
+import org.jetbrains.idea.maven.model.MavenPlugin;
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -19,18 +21,21 @@ public class VersionComparator {
     private static final String GROUP = "g";
     private static final String ARTIFACT = "a";
 
-    private final Map<String, String> moduleDependencies;
+    private final List<String> queryResults;
 
-    private final List<Dependency> mavenDependencies;
+    private List<Dependency> mavenDependencies;
 
-    public VersionComparator(Map<String, String> moduleDependencies, List<Dependency> mavenDependencies) {
-        this.moduleDependencies = moduleDependencies;
-        this.mavenDependencies = mavenDependencies;
+    private List<Plugin> mavenPlugins;
+
+    public VersionComparator(List<String> queryResults) {
+        this.queryResults = queryResults;
     }
 
-    public List<DependencyUpdateResult> compareVersions(List<String> jsons) {
+    public List<DependencyUpdateResult> compareDependencyVersions(Map<String, String> moduleDependencies, List<Dependency> mavenDependencies) {
+        this.mavenDependencies = mavenDependencies;
+
         List<DependencyUpdateResult> result = new ArrayList<>();
-        for (String json : jsons) {
+        for (String json : queryResults) {
             try {
                 JSONObject jsonObject = new JSONObject(json);
                 JSONObject response = jsonObject.getJSONObject(RESPONSE);
@@ -56,9 +61,49 @@ public class VersionComparator {
                     result.add(new DependencyUpdateResult(groupId, artifactId, currentVersion, latestVersion));
                 }
             } catch (JSONException je) {
-                // ignoring
+                // ignore
             }
         }
+        return result;
+    }
+
+    public List<DependencyUpdateResult> comparePluginVersions(List<MavenPlugin> projectPlugins, List<Plugin> mavenPlugins) {
+        this.mavenPlugins = mavenPlugins;
+
+        List<DependencyUpdateResult> result = new ArrayList<>();
+        for (String json : queryResults) {
+            try {
+                JSONObject jsonObject = new JSONObject(json);
+                JSONObject response = jsonObject.getJSONObject(RESPONSE);
+                JSONObject docsObject = response.getJSONArray(DOCS_ARRAY).getJSONObject(0);
+                String groupId = docsObject.getString(GROUP);
+                String artifactId = docsObject.getString(ARTIFACT);
+                String latestVersion = docsObject.getString(VERSION);
+                ComparableVersion latestVersionComparable = new ComparableVersion(latestVersion);
+
+                // if not managed by IntelliJ use current POM version instead
+                String currentVersion = null;
+                for (MavenPlugin projectPlugin : projectPlugins) {
+                    if (projectPlugin.getGroupId().equals(groupId) && projectPlugin.getArtifactId().equals(artifactId)) {
+                        currentVersion = projectPlugin.getVersion() != null ? projectPlugin.getVersion() : getCurrentMavenPluginVersion(groupId, artifactId);
+                    }
+                }
+
+                if (currentVersion == null) {
+                    // if still null skip version comparison
+                    continue;
+                }
+
+                ComparableVersion currentVersionComparable = new ComparableVersion(currentVersion);
+
+                if (latestVersionComparable.compareTo(currentVersionComparable) > 0) {
+                    result.add(new DependencyUpdateResult(groupId, artifactId, currentVersion, latestVersion));
+                }
+            } catch (JSONException je) {
+                // ignore
+            }
+        }
+
         return result;
     }
 
@@ -66,6 +111,15 @@ public class VersionComparator {
         for (Dependency mavenDependency : mavenDependencies) {
             if (mavenDependency.getGroupId().equals(groupId) && mavenDependency.getArtifactId().equals(artifactId)) {
                 return mavenDependency.getVersion();
+            }
+        }
+        return null;
+    }
+
+    private String getCurrentMavenPluginVersion(String groupId, String artifactId) {
+        for (Plugin mavenPlugin : mavenPlugins) {
+            if (mavenPlugin.getGroupId().equals(groupId) && mavenPlugin.getArtifactId().equals(artifactId)) {
+                return mavenPlugin.getVersion();
             }
         }
         return null;